requirements.txt
# Core dependencies
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=1.0.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Forecasting
prophet>=1.1.5

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0



